@@ -23,20 +23,20 @@
     "@uiw/react-color-shade-slider": "~1.0.10",
     "@uiw/react-color-sketch": "~1.0.10",
     "@uiw/react-color-wheel": "~1.0.10",
-    "react": "~17.0.2",
-    "react-dom": "~17.0.2",
-    "prismjs": "~1.27.0"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "prismjs": "^1.28.0"
   },
   "devDependencies": {
-    "@kkt/less-modules": "~7.1.1",
-    "@types/prismjs": "~1.26.0",
-    "@types/react": "~17.0.39",
-    "@types/react-dom": "~17.0.13",
-    "husky": "~7.0.4",
-    "lint-staged": "~12.3.4",
+    "@kkt/less-modules": "^7.2.0",
+    "@types/prismjs": "^1.26.0",
+    "@types/react": "^18.0.15",
+    "@types/react-dom": "^18.0.6",
+    "husky": "^8.0.1",
+    "lint-staged": "^13.0.3",
     "source-map-explorer": "~2.5.2",
-    "kkt": "~7.1.5",
-    "prettier": "~2.5.1"
+    "kkt": "^7.2.0",
+    "prettier": "^2.7.1"
   },
   "lint-staged": {
     "*.{js,jsx,tsx,ts,less,md,json}": [
